+++ django_project/frontend/src/pages/AnalysisResults/index.tsx
@@ -38,6 +38,7 @@ import { IoCloseSharp } from "react-icons/io5";
 import { ChevronUpIcon } from "@chakra-ui/icons";
 import Pagination from "../../components/Pagination";
 import ConfirmDeleteDialog from "../../components/ConfirmDeleteDialog";
+import { getAnalysisSummary } from "../../utils/analysisSummary";
 
 interface FeatureProperties {
   Project?: string;
@@ -55,19 +56,6 @@ interface AnalysisResults {
     latitude?: number;
     longitude?: number;
   };
-}
-
-interface AnalysisData {
-  analysis_results?: AnalysisResults;
-}
-
-interface AnalysisSummary {
-  title: string;
-  projectName: string;
-  locationName: string;
-  analysisType: string;
-  latitude?: number;
-  longitude?: number;
 }
 
 export default function AnalysisResults() {
@@ -138,33 +126,7 @@ export default function AnalysisResults() {
    }
   }, [analysisDeleted]);
 
-  const getAnalysisSummary = (analysis: any): AnalysisSummary => {
-    const { analysis_results } = analysis || {};
-    const { results, data } = analysis_results || {};
   
-    const features = results?.[0]?.features || [];
-    const { analysisType = "Analysis", latitude, longitude, landscape } = data || {};
-  
-    // Extract properties safely
-    const projectName = features?.[0]?.properties?.Project || "Unknown Project";
-  
-    // Construct a meaningful title
-    const title =
-      landscape && landscape !== "Unknown Landscape" 
-        ? `${analysisType} Analysis of ${landscape}`
-        : projectName === "Unknown Project"
-        ? `${analysisType} Results from Area`
-        : `${analysisType} Analysis of ${landscape} in the ${projectName} Landscape.`;
-    
-    return {
-      title,
-      projectName,
-      locationName: landscape,
-      analysisType,
-      latitude,
-      longitude,
-    };
-  };
   
 
   // Filtering based on search term
